requirements.txt
flask
torch
torchvision
numpy
pillow
matplotlib
h5py
scikit-learn
joblib
requests
openai
transformers
biopython
langchain
langchain-community
langchain-openai
langchain-huggingface
faiss-cpu
tqdm